changes model files to match new fields

+++ db/migrations/20180524172555-add-fear-to-articles.js
@@ -2,7 +2,7 @@
 
 module.exports = {
   up: (queryInterface, Sequelize) => {
-    return queryInterface.addColumn('Articles', 'fear', {type: Sequelize.integer});
+    return queryInterface.addColumn('Articles', 'fear', {type: Sequelize.FLOAT});
   },
 
   down: (queryInterface, Sequelize) => {
