+++ db/migrations/20180524172640-add-joy-to-articles.js
@@ -2,7 +2,7 @@
 
 module.exports = {
   up: (queryInterface, Sequelize) => {
-    return queryInterface.addColumn('Articles', 'joy', {type: Sequelize.integer});
+    return queryInterface.addColumn('Articles', 'joy', {type: Sequelize.FLOAT});
   },
 
   down: (queryInterface, Sequelize) => {
