+++ db/migrations/20180524172722-add-disgust-to-articles.js
@@ -2,7 +2,7 @@
 
 module.exports = {
   up: (queryInterface, Sequelize) => {
-    return queryInterface.addColumn('Articles', 'disgust', {type: Sequelize.integer});
+    return queryInterface.addColumn('Articles', 'disgust', {type: Sequelize.FLOAT});
   },
 
   down: (queryInterface, Sequelize) => {
