+++ db/migrations/20180524172943-add-sadness-to-events.js
@@ -2,7 +2,7 @@
 
 module.exports = {
   up: (queryInterface, Sequelize) => {
-    return queryInterface.addColumn('Events', 'sadness', {type: Sequelize.integer});
+    return queryInterface.addColumn('Events', 'sadness', {type: Sequelize.FLOAT});
   },
 
   down: (queryInterface, Sequelize) => {
