+++ db/migrations/20180524173022-add-joy-to-events.js
@@ -2,7 +2,7 @@
 
 module.exports = {
   up: (queryInterface, Sequelize) => {
-    return queryInterface.addColumn('Events', 'joy', {type: Sequelize.integer});
+    return queryInterface.addColumn('Events', 'joy', {type: Sequelize.FLOAT});
   },
 
   down: (queryInterface, Sequelize) => {
